move tests of worktable to tests folder
