+++ tests/worktable/tuple_primary_key.rs
@@ -1,0 +1,26 @@
+use worktable::prelude::*;
+use worktable::worktable;
+
+worktable! (
+    name: Test,
+    columns: {
+        id: u64 primary_key,
+        test: u64 primary_key,
+        another: i64,
+    }
+);
+
+#[test]
+fn insert() {
+    let table = TestWorkTable::default();
+    let row = TestRow {
+        id: 1,
+        test: 1,
+        another: 1,
+    };
+    let pk = table.insert(row.clone()).unwrap();
+    let selected_row = table.select(pk).unwrap();
+
+    assert_eq!(selected_row, row);
+    assert!(table.select((1, 0).into()).is_none())
+}
